+++ lib/node/pl-http/src/index.ts
@@ -1,23 +1,14 @@
-import { Agent, Client, Dispatcher, ProxyAgent } from 'undici';
-import CacheableLookup from 'cacheable-lookup';
-import { Resolver } from 'node:dns/promises';
+import { Agent, Client, Dispatcher, ProxyAgent, interceptors } from 'undici';
 
 export function defaultHttpDispatcher(httpProxy?: string): Dispatcher {
-  const cacheableLookup = new CacheableLookup({
-    resolver: new Resolver({ timeout: 3000, tries: 4 })
-  });
-
   const httpOptions: Client.Options = {
     allowH2: true,
     headersTimeout: 15e3,
     keepAliveTimeout: 15e3,
     keepAliveMaxTimeout: 60e3,
-    maxRedirections: 10,
-    connect: {
-      lookup: cacheableLookup.lookup.bind(cacheableLookup) as any
-    }
+    maxRedirections: 10
   };
 
-  if (httpProxy !== undefined) return new ProxyAgent({ uri: httpProxy, ...httpOptions });
-  else return new Agent(httpOptions);
+  const dispatcher = httpProxy !== undefined ? new ProxyAgent({ uri: httpProxy, ...httpOptions }) : new Agent(httpOptions);
+  return dispatcher; //.compose(interceptors.dns())
 }