+++ lib/node/pl-drivers/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @milaboratories/pl-drivers
 
+## 1.5.24
+
+### Patch Changes
+
+- @milaboratories/pl-client@2.7.11
+- @milaboratories/pl-tree@1.4.30
+
+## 1.5.23
+
+### Patch Changes
+
+- @milaboratories/pl-client@2.7.10
+- @milaboratories/pl-tree@1.4.29
+
+## 1.5.22
+
+### Patch Changes
+
+- Updated dependencies [2c75b72]
+  - @milaboratories/pl-client@2.7.9
+  - @milaboratories/pl-tree@1.4.28
+
 ## 1.5.21
 
 ### Patch Changes