+++ tools/block-tools/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @platforma-sdk/block-tools
 
+## 2.5.18
+
+### Patch Changes
+
+- Updated dependencies [3bf8838]
+  - @milaboratories/pl-http@1.1.0
+
+## 2.5.17
+
+### Patch Changes
+
+- Updated dependencies [e6ad278]
+  - @milaboratories/pl-http@1.0.7
+
 ## 2.5.16
 
 ### Patch Changes