+++ tools/r-builder/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @platforma-sdk/r-builder
 
+## 1.1.2
+
+### Patch Changes
+
+- b791b1e: Don't copy libraries twice to avoid issues with permissions (no 'w' for current user)
+
+## 1.1.1
+
+### Patch Changes
+
+- 23c98e3: Fix postinstall steps
+
+## 1.1.0
+
+### Minor Changes
+
+- f78c9a9: Support rocky linux
+
+## 1.0.0
+
+### Major Changes
+
+- 70a8665: Initial release
+
+## 0.0.5
+
+### Patch Changes
+
+- ef027e2: Fix download logic for CI (when out is not TTY)
+
+## 0.0.4
+
+### Patch Changes
+
+- a8ec363: Drop unused xz import
+
+## 0.0.3
+
+### Patch Changes
+
+- c3060c8: Polishing
+
 ## 0.0.2
 
 ### Patch Changes